Added createMatrix.m
Details of note:
1. This program resizes the image to 1024x768
2. I haven't seen the graphical image after the processing, so be sure
to check if you are able
3. The parameter is the file as a string
Ex. createMatrix('bear1.jpeg')
Other note. If we are to be generating 1 centroid per classification,
we're gonna run into issues if the images used to generate the centroid
are very different (different perspectives of the same thing = very
different). For your image choices, try to get similar
poses/perspectives on your objects